feat: externalize dashboard data fetching
