+++ package.json
@@ -6,20 +6,22 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "test": "vitest run"
   },
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "recharts": "^2.12.7",
-    "react-simple-maps": "^3.0.0"
+    "react-simple-maps": "^3.0.0",
+    "recharts": "^2.12.7"
   },
   "devDependencies": {
     "@types/node": "^22.14.0",
     "@types/react": "^18.2.24",
     "@types/react-dom": "^18.2.8",
+    "@vitejs/plugin-react": "^4.2.0",
     "typescript": "~5.8.2",
     "vite": "^5.2.0",
-    "@vitejs/plugin-react": "^4.2.0"
+    "vitest": "^3.2.4"
   }
 }
